# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "Astar token statistics API",
    "description": "Provides Astar networks statistic information."
  },
  "host": "api.astar.network",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/v3/{network}/dapps-staking/chaindapps": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves list of dapps (basic info) registered for dapps staking",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya",
            "enum": [
              "astar",
              "shiden",
              "shibuya"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v3/{network}/dapps-staking/stake-info/{address}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retrieves the amount of stake of participant",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar, shiden, shibuya",
            "enum": [
              "astar",
              "shiden",
              "shibuya"
            ]
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Participant address to get stats for"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v3/{network}/dapps-staking/get-period-range/{period}": {
      "get": {
        "tags": [
          "Dapps Staking"
        ],
        "description": "Retreives the start and end block numbers for the given dApp staking period.",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The network name. Supported networks: astar",
            "enum": [
              "astar",
              "shiden",
              "shibuya"
            ]
          },
          {
            "name": "period",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "dApp staking period."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}